requirements.txt
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
astroid==3.3.9 ; python_version >= "3.11" and python_version < "4.0"
backports-tarfile==1.2.0 ; python_version >= "3.11" and python_version < "3.12"
black==25.1.0 ; python_version >= "3.11" and python_version < "4.0"
build==1.2.2.post1 ; python_version >= "3.11" and python_version < "4.0"
cachecontrol[filecache]==0.14.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux" or platform_python_implementation == "PyPy")
cfgv==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows" or os_name == "nt")
crashtest==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
cryptography==44.0.2 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
dill==0.3.9 ; python_version >= "3.12" and python_version < "4.0" or python_version == "3.11"
distlib==0.3.9 ; python_version >= "3.11" and python_version < "4.0"
django-environ==0.12.0 ; python_version >= "3.11" and python_version < "4"
django-phonenumber-field[phonenumbers]==8.0.0 ; python_version >= "3.11" and python_version < "4.0"
django==5.2 ; python_version >= "3.11" and python_version < "4.0"
dulwich==0.22.8 ; python_version >= "3.11" and python_version < "4.0"
environ==1.0 ; python_version >= "3.11" and python_version < "4.0"
fastjsonschema==2.21.1 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.18.0 ; python_version >= "3.11" and python_version < "4.0"
findpython==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
flake8==7.2.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
identify==2.6.9 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.11" and python_version < "3.12"
installer==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
isort==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
jaraco-context==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
jaraco-functools==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
jeepney==0.9.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
keyring==25.6.0 ; python_version >= "3.11" and python_version < "4.0"
mccabe==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
more-itertools==10.6.0 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
pbs-installer[download,install]==2025.3.17 ; python_version >= "3.11" and python_version < "4.0"
phonenumbers==9.0.2 ; python_version >= "3.11" and python_version < "4.0"
pkginfo==1.12.1.2 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.3.7 ; python_version >= "3.11" and python_version < "4.0"
poetry-core==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
poetry-plugin-export==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
poetry==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
pre-commit==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.11" and python_version < "4.0"
pycodestyle==2.13.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux" or platform_python_implementation == "PyPy")
pyflakes==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
pylint-django==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
pylint-plugin-utils==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
pylint==3.3.6 ; python_version >= "3.11" and python_version < "4.0"
pyproject-hooks==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
rapidfuzz==3.13.0 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.3 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.11" and python_version < "4.0"
trove-classifiers==2025.3.19.19 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.1 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.30.0 ; python_version >= "3.11" and python_version < "4.0"
xattr==1.1.4 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "darwin"
yapf==0.43.0 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.11" and python_version < "3.12"
zstandard==0.23.0 ; python_version >= "3.11" and python_version < "4.0"
